requirements.txt
addict==2.4.0
albumentations==1.3.0
apex==0.9.10dev
basicsr==1.4.2
cv_bridge==1.13.0.post0
diffusers==0.27.2
einops==0.3.0
external==0.0.1
fairscale==0.4.13
ftfy==6.2.0
fvcore==0.1.5.post20221221
h5py==3.11.0
hydra-core==1.3.2
imutils==0.5.4
iopath==0.1.10
lmdb==1.4.1
matplotlib==3.7.5
mmdet==3.3.0
numpy==1.23.1
onnx==1.16.0
onnxruntime==1.17.3
open_clip_torch==2.0.2
opencv_python==4.9.0.80
openvino==2024.1.0
packaging==24.0
pandas==2.0.3
parrots==1.0.3
Pillow==10.3.0
prettytable==3.10.0
pycocotools==2.0.7
Pygments==2.18.0
pytorch_lightning==1.5.0
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.12.25
Requests==2.31.0
safetensors==0.4.3
scipy==1.9.1
setuptools==68.2.2
Shapely==2.0.4
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2023.7.10
timm==0.9.16
torch==2.2.2
torch_tb_profiler==0.4.3
torchvision==0.17.2
tqdm==4.66.2
transformers
turbojpeg==0.0.2
xformers==0.0.25.post1
yapf==0.32.0
